+++ src/main/java/org/folio/entlinks/service/links/InstanceAuthorityLinkingRulesService.java
@@ -1,7 +1,5 @@
 package org.folio.entlinks.service.links;
 
-import static org.folio.entlinks.config.constants.CacheNames.AUTHORITY_LINKING_RULES_CACHE;
-
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -9,8 +7,6 @@ import org.folio.entlinks.domain.entity.InstanceAuthorityLinkingRule;
 import org.folio.entlinks.domain.repository.LinkingRulesRepository;
 import org.folio.entlinks.exception.LinkingRuleNotFoundException;
 import org.folio.entlinks.service.links.validator.LinkingRuleValidator;
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,17 +19,11 @@ public class InstanceAuthorityLinkingRulesService {
   private final LinkingRulesRepository repository;
   private final LinkingRuleValidator validator;
 
-  @Cacheable(cacheNames = AUTHORITY_LINKING_RULES_CACHE,
-             key = "@folioExecutionContext.tenantId",
-             unless = "#result.isEmpty()")
   public List<InstanceAuthorityLinkingRule> getLinkingRules() {
     log.info("Loading linking rules");
     return repository.findAll(Sort.by("id").ascending());
   }
 
-  @Cacheable(cacheNames = AUTHORITY_LINKING_RULES_CACHE,
-             key = "@folioExecutionContext.tenantId + ':' + #authorityField",
-             unless = "#result.isEmpty()")
   public List<InstanceAuthorityLinkingRule> getLinkingRulesByAuthorityField(String authorityField) {
     log.info("Loading linking rules for [authorityField: {}]", authorityField);
     return repository.findByAuthorityField(authorityField);
@@ -46,7 +36,6 @@ public class InstanceAuthorityLinkingRulesService {
   }
 
   @Transactional
-  @CacheEvict(cacheNames = AUTHORITY_LINKING_RULES_CACHE, allEntries = true)
   public void updateLinkingRule(Integer ruleId, InstanceAuthorityLinkingRule linkingRulePatch) {
     log.info("Updating linking rule [ruleId: {}, change: {}]", ruleId, linkingRulePatch);
     var existingRule = getLinkingRule(ruleId);